fix num optiimzation steps for LR schedule. not fully tested yet
