+++ test/test_doc_classification.py
@@ -2,7 +2,7 @@ import logging
 
 from farm.data_handler.data_silo import DataSilo
 from farm.data_handler.processor import GermEval18CoarseProcessor
-from farm.experiment import initialize_optimizer
+from farm.modeling.optimization import initialize_optimizer
 from farm.infer import Inferencer
 from farm.modeling.adaptive_model import AdaptiveModel
 from farm.modeling.language_model import Bert
@@ -48,8 +48,7 @@ def test_doc_classification(caplog):
         model=model,
         learning_rate=2e-5,
         warmup_proportion=0.1,
-        n_examples=data_silo.n_samples("train"),
-        batch_size=batch_size,
+        n_batches=len(data_silo.loaders["train"]),
         n_epochs=1)
 
     trainer = Trainer(
@@ -73,5 +72,5 @@ def test_doc_classification(caplog):
     ]
     model = Inferencer.load(save_dir)
     result = model.run_inference(dicts=basic_texts)
-    assert result[0]["predictions"][0]["label"] == "OTHER"
-    assert abs(result[0]["predictions"][0]["probability"] - 0.5358161) <= 0.0001
+    assert result[0]["predictions"][0]["label"] == "OFFENSE"
+    assert abs(result[0]["predictions"][0]["probability"] - 0.5256448) <= 0.0001
