+++ test/test_question_answering.py
@@ -4,7 +4,7 @@ import pprint
 
 from farm.data_handler.data_silo import DataSilo
 from farm.data_handler.processor import SquadProcessor
-from farm.experiment import initialize_optimizer
+from farm.modeling.optimization import initialize_optimizer
 from farm.infer import Inferencer
 from farm.modeling.adaptive_model import AdaptiveModel
 from farm.modeling.language_model import Bert
@@ -51,8 +51,7 @@ def test_qa(caplog):
         model=model,
         learning_rate=1e-5,
         warmup_proportion=0.2,
-        n_examples=data_silo.n_samples("train"),
-        batch_size=batch_size,
+        n_batches=len(data_silo.loaders["train"]),
         n_epochs=n_epochs,
     )
     trainer = Trainer(
